feat: 잔돈 충전 탭의 보유 금액 및 동전 개수가 다른 탭 이동 후 다시 돌아와도 그대로 보여지는 기능 구현
잔돈 충전 탭에서 잔돈을 충전하고 다른 탭으로 이동 후 다시 잔돈 충전
탭으로 돌아왔을 때 현재 보유 금액 및 동전 개수가 보여지는 기능
현재 보유 금액 및 동전 개수를 출력하는 기능을 외부 모듈로 분리하고 잔돈
충전 탭을 클릭할 때 출력 모듈 호출
