+++ src/attachVendingMachineChargeEvent.js
@@ -1,4 +1,5 @@
 import { DOM_ID_SELECTOR, ERROR_MESSAGE } from './constants.js';
+import printVendingMachineAmountAndCoinTable from './printVendingMachineAmountAdnCoinTable.js';
 
 const isValidPrice = (price) => {
   const priceNumber = Number(price);
@@ -19,25 +20,6 @@ const clearVendingMachineChargeInput = () => {
   document.getElementById(DOM_ID_SELECTOR.vendingMachineChargeInput).value = '';
 };
 
-const printVendingMachineAmount = (amount) => {
-  const $vendingMachineChargeAmount = document.getElementById(DOM_ID_SELECTOR.vendingMachineChargeAmount);
-  $vendingMachineChargeAmount.innerText = amount;
-};
-
-const printVendingMachineCoinTable = (coinQuantity) => {
-  const $vendingMachineCoin500Quantity = document.getElementById(DOM_ID_SELECTOR.vendingMachineCoin500Quantity);
-  $vendingMachineCoin500Quantity.innerText = coinQuantity[500];
-
-  const $vendingMachineCoin100Quantity = document.getElementById(DOM_ID_SELECTOR.vendingMachineCoin100Quantity);
-  $vendingMachineCoin100Quantity.innerText = coinQuantity[100];
-
-  const $vendingMachineCoin50Quantity = document.getElementById(DOM_ID_SELECTOR.vendingMachineCoin50Quantity);
-  $vendingMachineCoin50Quantity.innerText = coinQuantity[50];
-
-  const $vendingMachineCoin10Quantity = document.getElementById(DOM_ID_SELECTOR.vendingMachineCoin10Quantity);
-  $vendingMachineCoin10Quantity.innerText = coinQuantity[10];
-};
-
 const attachVendingMachineChargeEvent = (coin) => {
   const $vendingMachineChargeButton = document.getElementById(DOM_ID_SELECTOR.vendingMachineChargeButton);
 
@@ -48,8 +30,7 @@ const attachVendingMachineChargeEvent = (coin) => {
     if (isValidPrice(price)) {
       clearVendingMachineChargeInput();
       coin.charge(Number(price));
-      printVendingMachineAmount(coin.getAmount());
-      printVendingMachineCoinTable(coin.getCoinQuantity());
+      printVendingMachineAmountAndCoinTable(coin);
     }
   });
 };
